chore(schema): seminar_schedules nullable 처리 + 도메인 파일 분리
- seminar_schedules의 description/started_at/ended_at/venue_name/notice nullable로 변경
  (시간/장소 미정 일정 표시 지원)
- src/schema/seminar.ts 신설: seminar 5개 테이블 + 5개 relations
- 단방향 import (seminar.ts → schema.ts) 채택
  양방향은 enum top-level 호출 시 ESM cycle TDZ 에러 발생
- members.seminarAttendanceRecords / generations.seminarSchedules 역참조 제거
  (현재 사용처 0건)
- (부수) 모든 테이블 timestamp default 정합성 보정
  defaultNow()가 0000 마이그레이션에 누락된 부분 회복
