+++ src/schema/index.ts
@@ -1,1 +1,2 @@
 export * from './schema';
+export * from './seminar';
